feat: changed internal map from Vec to boxed slice
BREAKING CHANGE: This will break old `new(...)` invocations. In fact,
now you must pass a boxed slice in the last parameter.

This should save the allocation of the Vector's capacity field. The map
cannot be resized anyway, so it is a net memory improvement (even if not
huge, unless you are loading **a lot** of maps at the same time).